added extended section page for show all
